paper-retrieval/requirements.txt
openai>=1.0.0
click>=8.0.0
pyyaml>=6.0
requests>=2.28.0
arxiv>=2.1.0
rich>=13.0.0
tenacity>=8.0.0
